+++ integration/templates/next-app-router/src/app/pricing-table/page.tsx
@@ -1,4 +1,4 @@
-import { PricingTable } from '@clerk/nextjs';
+import { PricingTable, Protect } from '@clerk/nextjs';
 
 export default async function PricingTablePage({
   searchParams,
@@ -6,5 +6,18 @@ export default async function PricingTablePage({
   searchParams: Promise<{ newSubscriptionRedirectUrl: string }>;
 }) {
   const newSubscriptionRedirectUrl = (await searchParams).newSubscriptionRedirectUrl;
-  return <PricingTable newSubscriptionRedirectUrl={newSubscriptionRedirectUrl} />;
+  return (
+    <>
+      <Protect plan='free_user'>
+        <p>user in free</p>
+      </Protect>
+      <Protect plan='pro'>
+        <p>user in pro</p>
+      </Protect>
+      <Protect plan='plus'>
+        <p>user in plus</p>
+      </Protect>
+      <PricingTable newSubscriptionRedirectUrl={newSubscriptionRedirectUrl} />
+    </>
+  );
 }