+++ packages/clerk-js/src/ui/components/Checkout/CheckoutPage.tsx
@@ -20,8 +20,9 @@ import { useCheckout } from './useCheckout';
 
 export const CheckoutPage = (props: __internal_CheckoutProps) => {
   const { translateError } = useLocalizations();
+  const { t } = useLocalizations();
   const { planId, planPeriod, subscriberType, onSubscriptionComplete } = props;
-  const { setIsOpen, isOpen } = useDrawerContext();
+  const { setIsOpen } = useDrawerContext();
   const prefersReducedMotion = usePrefersReducedMotion();
   const { animations: layoutAnimations } = useAppearance().parsedLayout;
   const isMotionSafe = !prefersReducedMotion && layoutAnimations === true;
@@ -44,10 +45,8 @@ export const CheckoutPage = (props: __internal_CheckoutProps) => {
   };
 
   useEffect(() => {
-    if (isOpen) {
-      revalidate();
-    }
-  }, [isOpen]);
+    return invalidate;
+  }, []);
 
   if (isLoading || plansLoading) {
     return (
@@ -129,11 +128,10 @@ export const CheckoutPage = (props: __internal_CheckoutProps) => {
             </LineItems.Root>
           </Box>
           <Box sx={t => ({ padding: t.space.$4 })}>
-            {/* TODO(@Commerce): needs localization */}
             <Alert
               variant='info'
               colorScheme='info'
-              title={`You cannot subscribe to this plan by paying monthly. To subscribe to this plan, you need to choose to pay annually.`}
+              title={localizationKeys('commerce.cannotSubscribeMonthly')}
             />
           </Box>
         </Flex>
@@ -156,7 +154,7 @@ export const CheckoutPage = (props: __internal_CheckoutProps) => {
           variant='danger'
           colorScheme='danger'
         >
-          {errors ? translateError(errors[0]) : 'There was a problem, please try again later.'}
+          {errors ? translateError(errors[0]) : t(localizationKeys('unstable__errors.form_param_value_invalid'))}
         </Alert>
       </Flex>
     </Drawer.Body>